test(*):move events check test from exchangeV2 to transfer-manager
