+++ exchange-v2/test/v2/ExchangeV2.rarible.test.js
@@ -648,180 +648,6 @@ contract("ExchangeV2, sellerFee + buyerFee =  6%,", accounts => {
 
 	})	//Do matchOrders(), orders dataType == V1, MultipleBeneficiary
 
-	describe("Catch emit event Transfer", () => {
-		it("From ETH(DataV1) to ERC721(DataV1) Protocol, check emit ", async () => {
-			const seller = accounts[1];
-			const sellerRoyaltiy = accounts[4];
-			const seller2 = accounts[3];
-			const buyer = accounts[2];
-			const originLeft1 = accounts[5];
-			const originLeft2 = accounts[6];
-			const originRight = accounts[7];
-
-			await erc721V1.mint(seller, erc721TokenId1, [[sellerRoyaltiy, 1000]]);
-    	await erc721V1.setApprovalForAll(transferProxy.address, true, {from: seller});
-
-			let addrOriginLeft = [[originLeft1, 500], [originLeft2, 600]];
- 			let addrOriginRight = [[originRight, 700]];
- 			let encDataLeft = await encDataV1([ [[buyer, 10000]], addrOriginLeft ]);
- 			let encDataRight = await encDataV1([ [[seller, 5000], [seller2, 5000]], addrOriginRight ]);
-
-			const left = Order(buyer, Asset(ETH, "0x", 200), ZERO, Asset(ERC721, enc(erc721V1.address, erc721TokenId1), 1), 1, 0, 0, ORDER_DATA_V1, encDataLeft);
-    	const right = Order(seller, Asset(ERC721, enc(erc721V1.address, erc721TokenId1), 1), ZERO, Asset(ETH, "0x", 200), 1, 0, 0, ORDER_DATA_V1, encDataRight);
-    	let signatureRight = await getSignature(right, seller);
-    	let tx = await testing.matchOrders(left, "0x", right, signatureRight, { from: buyer, value: 300, gasPrice: 0 });
-			let errorCounter = 0
-//			eventEmitted  - срабатывает по нескольким transfer, для фиксации ошибки нужно чтоб все трансферы завалились
-			truffleAssert.eventEmitted(tx, 'Transfer', (ev) => {
-				let result = false;
-				switch (ev.to){
-					case protocol:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != PROTOCOL)) {
-							console.log("Error in protocol check:");
-							errorCounter++;
-						}
-					break
-					case seller:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != PAYOUT) ) {
-							console.log("Error in seller check:");
-							errorCounter++;
-						}
-					break
-					case sellerRoyaltiy:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != ROYALTY) ) {
-							console.log("Error in seller check:");
-							errorCounter++;
-						}
-					break
-					case seller2:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != PAYOUT) ) {
-							console.log("Error in seller2 check:");
-							errorCounter++;
-						}
-					break
-					case originLeft1:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != ORIGIN) ) {
-							console.log("Error in originLeft1 check:");
-							errorCounter++;
-						}
-					break
-					case originLeft2:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != ORIGIN) ) {
-							console.log("Error in originLeft2 check:");
-							errorCounter++;
-						}
-					break
-					case originRight:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != ORIGIN) ) {
-							console.log("Error in originRight check:");
-							errorCounter++;
-						}
-					break
-					case buyer:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != PAYOUT) ){
-							console.log("Error in buyer check:");
-							errorCounter++;
-						}
-					break
-				}
-				if (errorCounter > 0) {
-					result = false;
-				} else {
-					result = true;
-				}
-				return result;
-    	}, "Transfer shuold be emietted with correct parameters ");
-			assert.equal(errorCounter, 0); //фиксируем наличие ошибок тут
-    })
-
-		it("From ERC1155(DataV2) to ETH(DataV1) Protocol, check emit ", async () => {
-			const seller = accounts[1];
-			const sellerRoyaltiy = accounts[4];
-			const seller2 = accounts[3];
-			const buyer = accounts[2];
-			const originLeft1 = accounts[5];
-			const originLeft2 = accounts[6];
-			const originRight = accounts[7];
-
-			await erc1155_v2.mint(seller, erc1155TokenId1, [[sellerRoyaltiy, 1000]], 10);
-    	await erc1155_v2.setApprovalForAll(transferProxy.address, true, {from: seller});
-
-			let addrOriginLeft = [[originLeft1, 500], [originLeft2, 600]];
- 			let addrOriginRight = [[originRight, 700]];
- 			let encDataLeft = await encDataV1([ [[seller, 5000], [seller2, 5000]] , addrOriginLeft ]);
- 			let encDataRight = await encDataV1([ [[buyer, 10000]], addrOriginRight ]);
-
-			const left = Order(seller, Asset(ERC1155, enc(erc1155_v2.address, erc1155TokenId1), 5), ZERO, Asset(ETH, "0x", 200), 1, 0, 0, ORDER_DATA_V1, encDataLeft);
-			const right = Order(buyer, Asset(ETH, "0x", 200), ZERO, Asset(ERC1155, enc(erc1155_v2.address, erc1155TokenId1), 5), 1, 0, 0, ORDER_DATA_V1, encDataRight);
-
-    	let signatureRight = await getSignature(right, buyer);
-    	let tx = await testing.matchOrders(left, "0x", right, signatureRight, { from: seller, value: 300, gasPrice: 0 });
-			let errorCounter = 0
-//			eventEmitted  - срабатывает по нескольким transfer, для фиксации ошибки нужно чтоб все трансферы завалились
-			truffleAssert.eventEmitted(tx, 'Transfer', (ev) => {
-				let result = false;
-				switch (ev.to){
-					case protocol:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != PROTOCOL)) {
-							console.log("Error in protocol check:");
-							errorCounter++;
-						}
-					break
-					case seller:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != PAYOUT) ) {
-							console.log("Error in seller check:");
-							errorCounter++;
-						}
-					break
-					case sellerRoyaltiy:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != ROYALTY) ) {
-							console.log("Error in seller check:");
-							errorCounter++;
-						}
-					break
-					case seller2:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != PAYOUT) ) {
-							console.log("Error in seller2 check:");
-							errorCounter++;
-						}
-					break
-					case originLeft1:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != ORIGIN) ) {
-							console.log("Error in originLeft1 check:");
-							errorCounter++;
-						}
-					break
-					case originLeft2:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != ORIGIN) ) {
-							console.log("Error in originLeft2 check:");
-							errorCounter++;
-						}
-					break
-					case originRight:
-						if ((ev.transferDirection != TO_MAKER) && (ev.transferType != ORIGIN) ) {
-							console.log("Error in originRight check:");
-							errorCounter++;
-						}
-					break
-					case buyer:
-						if ((ev.transferDirection != TO_TAKER) && (ev.transferType != PAYOUT) ){
-							console.log("Error in buyer check:");
-							errorCounter++;
-						}
-					break
-				}
-				if (errorCounter > 0) {
-					result = false;
-				} else {
-					result = true;
-				}
-				return result;
-    	}, "Transfer shuold be emietted with correct parameters ");
-			assert.equal(errorCounter, 0); //фиксируем наличие ошибок тут
-    })
-
-	}) //Catch emit event Transfer
-
 	describe("Exchange with Royalties", () => {
 		it("Royalties by owner, token 721 to ETH", async () => {
 			await erc721.mint(accounts[1], erc721TokenId1);
